requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.9.1
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.4.0
Pygments==2.3.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
pytz==2018.9
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
Werkzeug==0.14.1
whitenoise==4.1.2
wrapt==1.11.1
yapf==0.26.0
